fix locale management for devise
before, when a user tried to access a protected controller while unauthenticated, the locale would be lost and authenticate_user! would redirect him to the login page in the defualt locale

this has been fixed by only setting the locale if it is passed as a param (e.g. user wants to change locale), or if the locale hasn't been set yet